+++ recipes/libavrocpp/config.yml
@@ -1,3 +1,3 @@
 versions:
   1.10.1:
-    folder: 1.10.1
+    folder: all
